inf O-power cheat is now tested and available on all editions

+++ source/otherModifiers.h
@@ -4,6 +4,6 @@ void initOtherModifiers(int edition);
 
 void setMoney(u32 count);
 void setRemainingRepel(u32 count);
-void setRemainingOPower(u32 count);
+void setRemainingOPower(u8 count);
 
 void setHatchingStepCounter(u16 steps);
